[simi-common-utils] feat: Rename package name
