+++ simi-common/simi-common-utils/src/main/java/com/simi/labs/common/utils/SmpAssert.java
@@ -1,4 +1,4 @@
-package com.simi.common.util;
+package com.simi.labs.common.utils;
 
 
 import jakarta.annotation.Nullable;
