+++ simi-common/simi-common-utils/src/main/java/com/simi/labs/common/utils/data/StringUtil.java
@@ -1,4 +1,4 @@
-package com.simi.common.util.data;
+package com.simi.labs.common.utils.data;
 
 import lombok.experimental.UtilityClass;
 import org.apache.commons.lang3.StringUtils;
