+++ simi-common/simi-common-utils/src/main/java/com/simi/labs/common/utils/encode/AESUtil.java
@@ -1,4 +1,4 @@
-package com.simi.common.util.encode;
+package com.simi.labs.common.utils.encode;
 
 import lombok.experimental.UtilityClass;
