+++ simi-common/simi-common-utils/src/main/java/com/simi/labs/common/utils/log/StopWatchDebugger.java
@@ -1,4 +1,4 @@
-package com.simi.common.util.log;
+package com.simi.labs.common.utils.log;
 import jakarta.annotation.Nullable;
 import lombok.Data;
 import lombok.experimental.UtilityClass;
