+++ simi-common/simi-common-utils/src/main/java/com/simi/labs/common/utils/network/IpUtilsWebFlux.java
@@ -1,4 +1,4 @@
-package com.simi.common.util.network;
+package com.simi.labs.common.utils.network;
 
 import lombok.experimental.UtilityClass;
 import lombok.extern.slf4j.Slf4j;
